Add FFT and sampling at required frequency (broken!)
This code of this commit does not work properly!
The functions that manage sampling from the ADC do not sample at the
correct specified frequency. Also is unclear if the FFT it's working
correctly.

With this commit i also added all dependencies for the FFT module (they
are a lot of files).